+++ HISTORY.rst
@@ -144,6 +144,7 @@
 - Update `FormFieldMixinFilteredQueryset` to load content from
   `source_queryset' and not `queryset`.
 - Add file metadata driver for OpenAI.
+- Add `content` as a model attribute for the document file sandbox.
 
 4.9.5 (XXXX-XX-XX)
 ==================